src/camera/requirements/dev.txt
-r base.txt

aiortc-pyav-stub      == 0.1.1
opencv-contrib-python == 4.5.5.62
